requirements.txt
folium
Pillow
numpy
python-dateutil
opencv-python
selenium
scipy
geopy
matplotlib
google-cloud-storage
.